refactor: extract request context
