+++ server/src/Constructor.ts
@@ -1,0 +1,3 @@
+export type Constructor<T> = Function & {
+    prototype: T
+}
