+++ server/src/RequestId.ts
@@ -1,0 +1,1 @@
+export abstract class RequestId extends String {}
